yolo_cpu_requirements.txt
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aom=3.6.0=h6a678d5_0
blas=1.0=mkl
brotlicffi=1.0.9.2=py311h6a678d5_1
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.7.15=h06a4308_0
cairo=1.18.4=h44eff21_0
certifi=2025.8.3=py311h06a4308_0
cffi=1.17.1=py311h1fdaa30_1
charset-normalizer=3.3.2=pyhd3eb1b0_0
contourpy=1.3.3=pypi_0
cpuonly=2.0=0
cycler=0.12.1=pypi_0
dav1d=1.2.1=h5eee18b_0
expat=2.7.1=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.17.0=py311h06a4308_0
fontconfig=2.14.1=h55d465d_3
fonttools=4.59.1=pypi_0
freetype=2.13.3=h4a9f257_0
fribidi=1.0.10=h7b6447c_0
fsspec=2025.7.0=pypi_0
giflib=5.2.2=h5eee18b_0
gmp=6.3.0=h6a678d5_0
gmpy2=2.2.1=py311h5eee18b_0
gnutls=3.6.15=he1e5248_0
graphite2=1.3.14=h295c915_1
harfbuzz=10.2.0=hdfddeaa_1
icu=73.1=h6a678d5_0
idna=3.7=py311h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.6=py311h06a4308_0
jpeg=9e=h5eee18b_3
kiwisolver=1.4.9=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.16=h92b89f2_1
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h6a678d5_0
libavif=1.1.1=h5eee18b_0
libdeflate=1.22=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libglib=2.84.2=h37c7471_0
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.7.0=hde9077f_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h9f374a3_1
libwebp-base=1.3.2=h5eee18b_1
libxcb=1.17.0=h9b100fa_0
libxml2=2.13.8=hfdd30dd_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=3.0.2=py311h5eee18b_0
matplotlib=3.10.5=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py311h5eee18b_2
mkl_fft=1.3.11=py311h5eee18b_0
mkl_random=1.2.8=py311ha02d727_0
mpc=1.3.1=h5eee18b_0
mpfr=4.2.1=h5eee18b_0
mpmath=1.3.0=py311h06a4308_0
ncurses=6.5=h7934f7d_0
nettle=3.7.3=hbbd107a_1
networkx=3.4.2=py311h06a4308_0
numpy=2.0.1=py311h08b1b3b_1
numpy-base=2.0.1=py311hf175353_1
opencv-python=4.12.0.88=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=h0d4d230_1
openssl=3.0.17=h5eee18b_0
packaging=25.0=pypi_0
pandas=2.3.2=pypi_0
pcre2=10.42=hebb0a14_1
pillow=11.3.0=py311hb1c3d2d_0
pip=25.1=pyhc872135_2
pixman=0.46.4=h7934f7d_0
psutil=7.0.0=pypi_0
pthread-stubs=0.3=h0ce48e5_1
py-cpuinfo=9.0.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyparsing=3.2.3=pypi_0
pysocks=1.7.1=py311h06a4308_0
python=3.11.13=h1a3bd86_0
python-dateutil=2.9.0.post0=pypi_0
pytorch=2.5.1=py3.11_cpu_0
pytorch-mutex=1.0=cpu
pytz=2025.2=pypi_0
pyyaml=6.0.2=py311h5eee18b_0
readline=8.3=hc2a1206_0
requests=2.32.4=py311h06a4308_0
scipy=1.16.1=pypi_0
setuptools=78.1.1=py311h06a4308_0
six=1.17.0=pypi_0
sqlite=3.50.2=hb25bd0a_1
sympy=1.13.1=pypi_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.15=h54e0aa7_0
torchaudio=2.5.1=py311_cpu
torchvision=0.20.1=py311_cpu
typing_extensions=4.12.2=py311h06a4308_0
tzdata=2025.2=pypi_0
ultralytics=8.3.186=pypi_0
ultralytics-thop=2.0.16=pypi_0
urllib3=2.5.0=py311h06a4308_0
wheel=0.45.1=py311h06a4308_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-libxext=1.3.6=h9b100fa_0
xorg-libxrender=0.9.12=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.6.4=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0